requirements.txt
Django~=2.2
python-dotenv==0.14.0
django-mathfilters==1.0.0
